src/ClimateMonitor.IoT/RaspberrySensor/requirements.txt
adafruit-circuitpython-dht==4.0.4
requests==2.31.0
websockets==12.0
python-dateutil==2.9.0